+++ Source/EventFlow.TestHelpers/EventFlow.TestHelpers.csproj
@@ -68,6 +68,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Suites\TestSuiteForEventStore.cs" />
     <Compile Include="Suites\TestSuiteForReadModelStore.cs" />
+    <Compile Include="Suites\TestSuiteForScheduler.cs" />
     <Compile Include="Test.cs" />
     <Compile Include="TestsFor.cs" />
   </ItemGroup>
