+++ Source/EventFlow.TestHelpers/IntegrationTest.cs
@@ -20,7 +20,8 @@
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// 
+//
+
 using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Configuration;
@@ -47,8 +48,7 @@ namespace EventFlow.TestHelpers
         public void SetUpIntegrationTest()
         {
             var eventFlowOptions = EventFlowOptions.New
-                .AddEvents(EventFlowTestHelpers.Assembly)
-                .AddCommandHandlers(EventFlowTestHelpers.Assembly);
+                .AddDefaults(EventFlowTestHelpers.Assembly);
 
             Resolver = CreateRootResolver(eventFlowOptions);
